requirements.txt
kaggle
kagglehub
Pillow
-f https://download.pytorch.org/whl/nightly/cpu
torch
torchvision
torchaudio
scikit-learn
streamlit
captum
matplotlib
grad-cam